requirements.txt
Django==5.1.4
django-rest-framework==0.1.0
djangorestframework==3.15.2
django-filter==25.1
drf-spectacular==0.28.0
drf-yasg==1.21.8
numpy==2.2.0
pandas==2.2.3
